requirements.txt
python-dotenv
requests
gtts
transformers
torch
moviepy==1.0.3
imageio
imageio-ffmpeg
Pillow
tqdm
